requirements.txt
pandas
openpyxl
xlsxwriter
matplotlib
python-dotenv
openai
requests 